Moved calculations folder into /client/src to use in client and use jest for testing
